implement message store

+++ iNear/BadgeView.h
@@ -10,6 +10,6 @@
 
 @interface BadgeView : UIImageView
 
-- (void)setCount:(int)count;
+- (void)setCount:(NSUInteger)count;
 
 @end
